+++ src/decl/case-when-decl.h
@@ -1,0 +1,7 @@
+static
+r_ssize compute_size(
+  struct optional_r_ssize size,
+  r_obj* cases,
+  r_ssize cases_size,
+  struct r_lazy call
+);