+++ src/globals.c
@@ -49,11 +49,13 @@ void vctrs_init_globals(r_obj* ns) {
 
   // Symbols -----------------------------------------------------------
   syms.arg = r_sym("arg");
+  syms.cases_arg = r_sym("cases_arg");
   syms.default_arg = r_sym("default_arg");
   syms.dot_arg = r_sym(".arg");
   syms.dot_call = r_sym(".call");
   syms.dot_error_arg = r_sym(".error_arg");
   syms.dot_error_call = r_sym(".error_call");
+  syms.from_arg = r_sym("from_arg");
   syms.haystack_arg = r_sym("haystack_arg");
   syms.needles_arg = r_sym("needles_arg");
   syms.recurse = r_sym("recurse");
@@ -61,6 +63,7 @@ void vctrs_init_globals(r_obj* ns) {
   syms.times_arg = r_sym("times_arg");
   syms.to_arg = r_sym("to_arg");
   syms.value_arg = r_sym("value_arg");
+  syms.values_arg = r_sym("values_arg");
   syms.x_arg = r_sym("x_arg");
   syms.y_arg = r_sym("y_arg");